set up search results page
